+++ example/lib/htkc_theme_configurator.dart
@@ -1,4 +1,4 @@
-import 'package:example/lib/htkc_theme_color_selector.dart';
+import 'package:example/htkc_theme_color_selector.dart';
 import 'package:htkc_utils/htkc_utils.dart';
 
 class ThemeConfigurator extends StatelessWidget {