+++ example/lib/upgrade/main-appcast.dart
@@ -18,8 +18,8 @@ class MyApp extends StatelessWidget {
   MyApp({Key? key}) : super(key: key);
 
   static const appcastURL =
-      'https://raw.githubusercontent.com/larryaasen/upgrader/master/test/testappcast.xml';
-  final upgrader = HcUpgradeNewVersion(
+      'https://raw.githubusercontent.com/larryaasen/upgrade/master/test/testappcast.xml';
+  final upgrade = HcUpgradeNewVersion(
       appCastConfig:
           HcAppCastConfiguration(url: appcastURL, supportedOS: ['android']));
 
@@ -30,7 +30,7 @@ class MyApp extends StatelessWidget {
       home: Scaffold(
           appBar: AppBar(title: Text('Upgrader Appcast Example')),
           body: HcUpgradeAlert(
-            hcUpgrade: upgrader,
+            hcUpgrade: upgrade,
             child: Center(child: Text('Checking...')),
           )),
     );