+++ example/lib/upgrade/main-min-app-version.dart
@@ -18,8 +18,8 @@ class MyApp extends StatelessWidget {
   MyApp({Key? key}) : super(key: key);
 
   static const appcastURL =
-      'https://raw.githubusercontent.com/larryaasen/upgrader/master/test/testappcast.xml';
-  final upgrader = HcUpgradeNewVersion(
+      'https://raw.githubusercontent.com/larryaasen/upgrade/master/test/testappcast.xml';
+  final upgrade = HcUpgradeNewVersion(
     appCastConfig:
         HcAppCastConfiguration(url: appcastURL, supportedOS: ['android']),
     debugLogging: true,
@@ -33,7 +33,7 @@ class MyApp extends StatelessWidget {
       home: Scaffold(
           appBar: AppBar(title: Text('Upgrader Example')),
           body: HcUpgradeAlert(
-            hcUpgrade: upgrader,
+            hcUpgrade: upgrade,
             child: Center(child: Text('Checking...')),
           )),
     );