+++ lib/upgrade/hc_upgrade_new_version_alert.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:htkc_utils/hc_utils/hc_app_text_style.dart';
 import 'package:htkc_utils/upgrade/hc_upgrade_new_version.dart';
 import 'package:htkc_utils/upgrade/hc_upgrade_new_version_messages.dart';
 
@@ -22,19 +23,32 @@ class HcUpgradeAlert extends StatefulWidget {
     this.onUpdate,
     this.shouldPopScope,
     this.showIgnore = true,
+    this.buttonSize = 30,
+    this.buttonFontSize = 15,
+    this.primaryColor = Colors.green,
+    this.ignoreColor = Colors.orange,
+    this.laterColor = Colors.indigo,
+    this.buttonBorderColor = Colors.indigo,
     this.showLater = true,
     this.showReleaseNotes = true,
     this.cupertinoButtonTextStyle,
     this.dialogKey,
     this.navigatorKey,
     this.child,
-  }) : upgrader = hcUpgrade ?? HcUpgradeNewVersion.sharedInstance;
+  }) : upgrade = hcUpgrade ?? HcUpgradeNewVersion.sharedInstance;
 
-  /// The upgraders used to configure the upgrade dialog.
-  final HcUpgradeNewVersion upgrader;
+  /// The upgrade used to configure the upgrade dialog.
+  final HcUpgradeNewVersion upgrade;
 
   /// Can alert dialog be dismissed on tap outside of the alert dialog. Not used by [UpgradeCard]. (default: false)
   final bool canDismissDialog;
+
+  final Color primaryColor;
+  final Color laterColor;
+  final Color ignoreColor;
+  final Color buttonBorderColor;
+  final double buttonSize;
+  final double buttonFontSize;
 
   /// The upgrade dialog style. Used only on UpgradeAlert. (default: material)
   final HcUpgradeDialogStyle dialogStyle;
@@ -89,28 +103,28 @@ class HcUpgradeAlertState extends State<HcUpgradeAlert> {
   @override
   void initState() {
     super.initState();
-    widget.upgrader.initialize();
+    widget.upgrade.initialize();
   }
 
   /// Describes the part of the user interface represented by this widget.
   @override
   Widget build(BuildContext context) {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrade.debugLogging) {
       if (kDebugMode) {
         print('hcUpgrade: build UpgradeAlert');
       }
     }
 
     return StreamBuilder(
-      initialData: widget.upgrader.evaluationReady,
-      stream: widget.upgrader.evaluationStream,
+      initialData: widget.upgrade.evaluationReady,
+      stream: widget.upgrade.evaluationStream,
       builder:
           (BuildContext context, AsyncSnapshot<HcUpgradeEvaluateNeed> snapshot) {
         if ((snapshot.connectionState == ConnectionState.waiting ||
-                snapshot.connectionState == ConnectionState.active) &&
+            snapshot.connectionState == ConnectionState.active) &&
             snapshot.data != null &&
             snapshot.data!) {
-          if (widget.upgrader.debugLogging) {
+          if (widget.upgrade.debugLogging) {
             if (kDebugMode) {
               print("hcUpgrade: need to evaluate version");
             }
@@ -118,7 +132,7 @@ class HcUpgradeAlertState extends State<HcUpgradeAlert> {
 
           if (!displayed) {
             final checkContext = widget.navigatorKey != null &&
-                    widget.navigatorKey!.currentContext != null
+                widget.navigatorKey!.currentContext != null
                 ? widget.navigatorKey!.currentContext!
                 : context;
             checkVersion(context: checkContext);
@@ -132,24 +146,24 @@ class HcUpgradeAlertState extends State<HcUpgradeAlert> {
   /// Will show the alert dialog when it should be dispalyed.
   /// Only called by [HcUpgradeAlert] and not used by [UpgradeCard].
   void checkVersion({required BuildContext context}) {
-    final shouldDisplay = widget.upgrader.shouldDisplayUpgrade();
-    if (widget.upgrader.debugLogging) {
+    final shouldDisplay = widget.upgrade.shouldDisplayUpgrade();
+    if (widget.upgrade.debugLogging) {
       if (kDebugMode) {
         print('hcUpgrade: shouldDisplayReleaseNotes: shouldDisplayReleaseNotes');
       }
     }
     if (shouldDisplay) {
       displayed = true;
-      final appMessages = widget.upgrader.determineMessages(context);
+      final appMessages = widget.upgrade.determineMessages(context);
 
       Future.delayed(const Duration(milliseconds: 0), () {
         showTheDialog(
-          key: widget.dialogKey ?? const Key('upgrader_alert_dialog'),
+          key: widget.dialogKey ?? const Key('upgrade_alert_dialog'),
           context: context,
           title: appMessages.message(HcUpgradeMessage.title),
-          message: widget.upgrader.body(appMessages),
+          message: widget.upgrade.body(appMessages),
           releaseNotes:
-              shouldDisplayReleaseNotes ? widget.upgrader.releaseNotes : null,
+          shouldDisplayReleaseNotes ? widget.upgrade.releaseNotes : null,
           canDismissDialog: widget.canDismissDialog,
           messages: appMessages,
         );
@@ -158,7 +172,7 @@ class HcUpgradeAlertState extends State<HcUpgradeAlert> {
   }
 
   void onUserIgnored(BuildContext context, bool shouldPop) {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrade.debugLogging) {
       if (kDebugMode) {
         print('hcUpgrade: button tapped: ignore');
       }
@@ -168,7 +182,7 @@ class HcUpgradeAlertState extends State<HcUpgradeAlert> {
     final doProcess = widget.onIgnore?.call() ?? true;
 
     if (doProcess) {
-      widget.upgrader.saveIgnored();
+      widget.upgrade.saveIgnored();
     }
 
     if (shouldPop) {
@@ -177,7 +191,7 @@ class HcUpgradeAlertState extends State<HcUpgradeAlert> {
   }
 
   void onUserLater(BuildContext context, bool shouldPop) {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrade.debugLogging) {
       if (kDebugMode) {
         print('hcUpgrade: button tapped: later');
       }
@@ -192,7 +206,7 @@ class HcUpgradeAlertState extends State<HcUpgradeAlert> {
   }
 
   void onUserUpdated(BuildContext context, bool shouldPop) {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrade.debugLogging) {
       if (kDebugMode) {
         print('hcUpgrade: button tapped: update now');
       }
@@ -202,7 +216,7 @@ class HcUpgradeAlertState extends State<HcUpgradeAlert> {
     final doProcess = widget.onUpdate?.call() ?? true;
 
     if (doProcess) {
-      widget.upgrader.sendUserToAppStore();
+      widget.upgrade.sendUserToAppStore();
     }
 
     if (shouldPop) {
@@ -217,7 +231,7 @@ class HcUpgradeAlertState extends State<HcUpgradeAlert> {
 
   bool get shouldDisplayReleaseNotes =>
       widget.showReleaseNotes &&
-      (widget.upgrader.releaseNotes?.isNotEmpty ?? false);
+          (widget.upgrade.releaseNotes?.isNotEmpty ?? false);
 
   /// Show the alert dialog.
   void showTheDialog({
@@ -229,7 +243,7 @@ class HcUpgradeAlertState extends State<HcUpgradeAlert> {
     required bool canDismissDialog,
     required HcUpgradeMessages messages,
   }) {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrade.debugLogging) {
       if (kDebugMode) {
         print('hcUpgrade: showTheDialog title: $title');
         print('hcUpgrade: showTheDialog message: $message');
@@ -238,7 +252,7 @@ class HcUpgradeAlertState extends State<HcUpgradeAlert> {
     }
 
     // Save the date/time as the last time alerted.
-    widget.upgrader.saveLastAlerted();
+    widget.upgrade.saveLastAlerted();
 
     showDialog(
       barrierDismissible: canDismissDialog,
@@ -263,14 +277,14 @@ class HcUpgradeAlertState extends State<HcUpgradeAlert> {
   /// is false. Also called when the back button is pressed. Return true for
   /// the screen to be popped. Defaults to false.
   bool onWillPop() {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrade.debugLogging) {
       if (kDebugMode) {
         print('hcUpgrade: onWillPop called');
       }
     }
     if (widget.shouldPopScope != null) {
       final should = widget.shouldPopScope!();
-      if (widget.upgrader.debugLogging) {
+      if (widget.upgrade.debugLogging) {
         if (kDebugMode) {
           print('hcUpgrade: shouldPopScope=$should');
         }
@@ -291,7 +305,7 @@ class HcUpgradeAlertState extends State<HcUpgradeAlert> {
       HcUpgradeMessages messages) {
     // If installed version is below minimum app version, or is a critical update,
     // disable ignore and later buttons.
-    final isBlocked = widget.upgrader.blocked();
+    final isBlocked = widget.upgrade.blocked();
     final showIgnore = isBlocked ? false : widget.showIgnore;
     final showLater = isBlocked ? false : widget.showLater;
 
@@ -311,47 +325,82 @@ class HcUpgradeAlertState extends State<HcUpgradeAlert> {
             ],
           ));
     }
-    final textTitle = Text(title, key: const Key('upgrader.dialog.title'));
+    final textTitle = Column(
+      children: [
+        Text(title, key: const Key('upgrade.dialog.title'), style: HcAppTextStyle.boldTextStyle(size: 20),),
+        Divider(color: widget.buttonBorderColor,)
+      ],
+    );
     final content = Container(
         constraints: const BoxConstraints(maxHeight: 400),
         child: SingleChildScrollView(
             child: Column(
-          crossAxisAlignment:
+              crossAxisAlignment:
               cupertino ? CrossAxisAlignment.center : CrossAxisAlignment.start,
-          mainAxisSize: MainAxisSize.min,
-          children: <Widget>[
-            Text(message),
-            Padding(
-                padding: const EdgeInsets.only(top: 15.0),
-                child: Text(messages.message(HcUpgradeMessage.prompt) ?? '')),
-            if (notes != null) notes,
-          ],
-        )));
+              mainAxisSize: MainAxisSize.min,
+              children: <Widget>[
+                Text(message),
+                Padding(
+                    padding: const EdgeInsets.only(top: 15.0),
+                    child: Text(messages.message(HcUpgradeMessage.prompt) ?? '')),
+                if (notes != null) notes,
+              ],
+            )));
     final actions = <Widget>[
-      if (showIgnore)
-        button(cupertino, messages.message(HcUpgradeMessage.buttonTitleIgnore),
-            context, () => onUserIgnored(context, true)),
-      if (showLater)
-        button(cupertino, messages.message(HcUpgradeMessage.buttonTitleLater),
-            context, () => onUserLater(context, true)),
-      button(cupertino, messages.message(HcUpgradeMessage.buttonTitleUpdate),
-          context, () => onUserUpdated(context, !widget.upgrader.blocked())),
+      Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [
+          if (showIgnore)
+            button(cupertino, messages.message(HcUpgradeMessage.buttonTitleIgnore),
+                context, () => onUserIgnored(context, true), widget.ignoreColor),
+          if (showLater)
+            button(cupertino, messages.message(HcUpgradeMessage.buttonTitleLater),
+                context, () => onUserLater(context, true), widget.laterColor),
+          button(cupertino, messages.message(HcUpgradeMessage.buttonTitleUpdate),
+              context, () => onUserUpdated(context, !widget.upgrade.blocked()), widget.primaryColor),
+        ],
+      )
     ];
 
     return cupertino
         ? CupertinoAlertDialog(
-            key: key, title: textTitle, content: content, actions: actions)
+        key: key, title: textTitle, content: content, actions: actions)
         : AlertDialog(
-            key: key, title: textTitle, content: content, actions: actions);
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(20),
+        ),
+        key: key, title: textTitle, content: content, actions: actions);
   }
 
   Widget button(bool cupertino, String? text, BuildContext context,
-      VoidCallback? onPressed) {
-    return cupertino
+      VoidCallback? onPressed, Color color) {
+        return cupertino
         ? CupertinoDialogAction(
             textStyle: widget.cupertinoButtonTextStyle,
             onPressed: onPressed,
             child: Text(text ?? ''))
-        : TextButton(onPressed: onPressed, child: Text(text ?? ''));
+        : Expanded(
+            child: TextButton(onPressed: onPressed, child: Container(
+                height: widget.buttonSize,
+                decoration: BoxDecoration(
+                  border: Border.all(
+                    color: widget.buttonBorderColor,
+                  ),
+                  color: color,
+                  borderRadius: BorderRadius.circular(50),
+                ),
+                child: Center(
+                  child: Text(
+                    text??'',
+                    style: TextStyle(
+                      color: Colors.white,
+                      fontWeight: FontWeight.bold,
+                      fontSize: widget.buttonFontSize
+                    ),
+                  ),
+                ),
+              ),
+          ),
+        );
   }
 }